server/requirements.txt
requests
beautifulsoup4
sentence-transformers
tqdm
faiss-cpu
numpy